update default header.
remove rb header because I don't write ruby these days.
remove the encoding section because currently I'm working in an ascii
environment, this declaration in Python 2 is not necessary. Moreover,
in Python 3, the default setting is utf8 so there's no need to declare
it unless I'm writing Python code in GB18030 or something.

Signed-off-by: Kai Xia <xiaket@gmail.com>